Added configuration based spawn restrictions.
Add an entity name, and an array of blocks they are permitted to spawn
on. If left blank, the mob will not naturally spawn. If not included,
normal spawn mechanics will be used. If included, and blocks are
specified, the entity will ONLY naturally spawn if the block under them
is in the specified list.

Configuration comment includes a sample to restrict slimes.